+++ duoStrength.js
@@ -67,7 +67,6 @@ const CROWN_TOTAL_SELECTOR = "._3nYQm._1B0kf";
 const PRACTICE_TYPE_SELECT_MESSAGE_SELECTOR = ".aUkqy";
 const SKILL_TREE_SELECTOR = "._3YEom";
 const TIPS_PAGE_BODY_SELECTOR = "._1yyg2";
-const MOBILE_TIPS_PAGE_HEADER_SELECTOR = "._36P0W";
 const LOCKED_SKILL_POPOUT = "_1fMEX"; // used only in styles/stylesheet.css
 const CARTOON_CONTAINER = "F2B9m"; // used only in styles/stylesheet.css
 const HINT_SENTENCE_CONTAINER = "_1Q4WV"; // used only in styles/stylesheet.css
@@ -78,6 +77,7 @@ const PROGRESS_QUIZ_BOX = "_9GzaQ";
 const DUOLINGO_SCORE_BOX = "_37iXd";
 const FRIENDS_TABLE_TITLE_SELECTOR = ".-AHpg";
 const SOCIAL_BUTTONS_HEADER_SELECTOR = "._3qTe3";
+const NAVS_CONTAINER_SELECTOR = "._3ejvc";
 
 const flagYValues =
 {
@@ -677,7 +677,7 @@ function hasMetGoal()
 		(entry) =>
 		{
 			const day = (new Date(entry.datetime)).setHours(0,0,0,0);
-			return day == currentDate;
+			return day === currentDate;
 		}
 	);
 	
@@ -698,7 +698,7 @@ function currentProgress()
 	let lessonsToNextTreeLevel = 0;
 	for (let skill of skills)
 	{
-		if (skill.skill_progress.level == treeLevel)
+		if (skill.skill_progress.level === treeLevel)
 		{
 			lessonsToNextTreeLevel += skill.num_sessions_for_level - skill.level_sessions_finished;
 		}
@@ -713,7 +713,7 @@ function nextCheckpointIndex()
 	const firstLockedIndexReducer = (value, element, index) =>
 	{
 		const locked = element.querySelectorAll(`[src$="locked.svg"]`).length != 0;
-		if (value == -1 && locked)
+		if (value === -1 && locked)
 			return index;
 		else
 			return value;
@@ -725,7 +725,7 @@ function nextCheckpointIndex()
 function lessonsToNextCheckpoint()
 {
 	let index = nextCheckpointIndex();
-	if (index == -1)
+	if (index === -1)
 	{
 		return -1;
 	}
@@ -741,7 +741,7 @@ function lessonsToNextCheckpoint()
 			if (type != "checkpoint-badge")
 			{
 				// element is not a checkpoint
-				if (bonusSkillRow == null || !bonusSkillRow.contains(element))
+				if (bonusSkillRow === null || !bonusSkillRow.contains(element))
 					return idx < index;
 			}
 		}
@@ -749,14 +749,14 @@ function lessonsToNextCheckpoint()
 	const level0SkillsBeforeCheckpoint = skillsBeforeCheckpoint.filter(
 		(element) =>
 		{
-			return element.querySelectorAll(`[data-test="level-crown"]`).length == 0;
+			return element.querySelectorAll(`[data-test="level-crown"]`).length === 0;
 		}
 	);
 	return level0SkillsBeforeCheckpoint.reduce(
 		(total, element) =>
 		{
 			const skillTitle = element.querySelector(SKILL_NAME_SELECTOR).textContent;
-			const lessons = userData.language_data[languageCode].skills.find(skill => skill.short == skillTitle).missing_lessons;
+			const lessons = userData.language_data[languageCode].skills.find(skill => skill.short === skillTitle).missing_lessons;
 			return total + lessons;
 		}, 0
 	);
@@ -768,7 +768,7 @@ function progressEnds()
 	let lastDate = progress[endIndex][0];
 	const today = (new Date()).setHours(0,0,0,0);
 	
-	while (!hasMetGoal() && lastDate == today)
+	while (!hasMetGoal() && lastDate === today)
 	{
 		lastDate = progress[--endIndex][0];
 	}
@@ -876,7 +876,7 @@ function currentLanguageHistory()
 {
 	return userData.language_data[languageCode].calendar.filter(
 		(lesson) => {
-			return userData.language_data[languageCode].skills.find(skill => skill.id == lesson.skill_id) != null;
+			return userData.language_data[languageCode].skills.find(skill => skill.id === lesson.skill_id) != null;
 		}
 	);
 }
@@ -886,13 +886,13 @@ function daysToNextXPLevel(history, XPLeft)
 	const currentDate = (new Date()).setHours(0,0,0,0);
 	const metGoal = hasMetGoal();
 
-	if (history.length == 0)
+	if (history.length === 0)
 		return -1;
 
 	let XPTotal = 0;
 	
 	const firstDate = (new Date(history[0].datetime)).setHours(0,0,0,0);
-	if (firstDate == currentDate && !metGoal)
+	if (firstDate === currentDate && !metGoal)
 	{
 		// The only data from this language is for today and the goal has yet to me met.
 		return {time: -1};
@@ -956,7 +956,7 @@ function daysToNextTreeLevelByCalendar()
 
 	const calendar = currentLanguageHistory();
 
-	if (calendar.length == 0)
+	if (calendar.length === 0)
 	{
 		return {time: -1};
 	}
@@ -986,7 +986,7 @@ function daysToNextTreeLevelByCalendar()
 
 	const numLessons = practiceTimes.length;
 	
-	if (numLessons == 0) // possible if only calendar entries for this language are from today, and we have't met our goal
+	if (numLessons === 0) // possible if only calendar entries for this language are from today, and we have't met our goal
 	{
 		return {time: -1};
 	}
@@ -1044,7 +1044,7 @@ function daysToNextCheckpointByCalendar()
 {
 	const calendar = currentLanguageHistory();
 	
-	if (calendar.length == 0)
+	if (calendar.length === 0)
 	{
 		return {time: -1};
 	}
@@ -1319,7 +1319,7 @@ function graphSVG(data, ratio=1.5)
 		let title = document.createElementNS(svgns, "title");
 		title.textContent = `${data[i]} tree level contributing lesson${data[i]!=1?"s":""}`;
 
-		if (data[i] == 0)
+		if (data[i] === 0)
 			point.setAttributeNS(null, "fill", "white");
 
 		point.appendChild(title);
@@ -1341,8 +1341,8 @@ async function openLastSkillPopout()
 
 	if (skillName != null)
 	{
-		const skillNameElement = Array.from(document.querySelectorAll(SKILL_NAME_SELECTOR)).find(elem => elem.textContent == skillName);
-		if (skillNameElement == null)
+		const skillNameElement = Array.from(document.querySelectorAll(SKILL_NAME_SELECTOR)).find(elem => elem.textContent === skillName);
+		if (skillNameElement === null)
 			return false;
 
 		skillNameElement.click();
@@ -1357,7 +1357,7 @@ async function openLastSkillPopout()
 		else
 		{
 			const goldenOwlCheckpoint = document.querySelector(GOLDEN_OWL_CHECKPOINT_SELECTOR);
-			if (goldenOwlCheckpoint == null)
+			if (goldenOwlCheckpoint === null)
 				return false;
 
 			goldenOwlCheckpoint.click();
@@ -1427,7 +1427,7 @@ function addFlagBorders()
 						flag1.setAttribute("data-tree-level", treeLevel);
                         flag1.setAttribute("data-original-viewBox", `${x1} ${y1} ${width1} ${height1}`);
                         flag1.setAttribute("viewBox", `${+x1 + 5} ${+y1 + 5} ${+width1 - 10} ${+height1 - 10}`);
-						if (treeLevel == 5)
+						if (treeLevel === 5)
 						{
 							const crown = document.createElement("IMG");
 							crown.src = imgSrcBaseUrl+"/juicy-crown.svg";
@@ -1468,8 +1468,8 @@ function addStrengthBars(strengths)
 		let elementContents;
 
 		if (
-			skillElements[i].getAttribute("data-test") == "skill" ||
-			skillElements[i].getAttribute("data-test") == "intro-lesson"
+			skillElements[i].getAttribute("data-test") === "skill" ||
+			skillElements[i].getAttribute("data-test") === "intro-lesson"
 		)
 		{
 			elementContents = [
@@ -1490,7 +1490,7 @@ function addStrengthBars(strengths)
 		// name is a span element, normally it is the last element but if the skill is clicked then a new div is created with the start lesson button etc below the name plate. So need to find the correct span element.
 		for (let spanElement of Array.from(skillElements[i].childNodes[0].getElementsByTagName('span')))
 		{
-			if (spanElement.parentNode == elementContents[0].parentNode)
+			if (spanElement.parentNode === elementContents[0].parentNode)
 			{
 				elementContents.push(spanElement);
 			}
@@ -1514,7 +1514,7 @@ function addStrengthBars(strengths)
 				//
 				// Note this all assumes every tree that has bonus skills, has two.
 				
-				const bonusIndex = (strengths[1][0][0] == 0)? 1 : 0;
+				const bonusIndex = (strengths[1][0][0] === 0)? 1 : 0;
 				elementContents = elementContents.concat(strengths[1][bonusIndex]);
 				bonusElementsCount = 1;
 			}
@@ -1865,7 +1865,7 @@ function displayNeedsStrengthening(needsStrengthening, cracked = false, needsSor
 			let bonusSkillIndex = i - needsStrengthening[0].length;
 			const skill = needsStrengthening[1][bonusSkillIndex];
 
-			skillLink.href = `/skill/${languageCode}/${skill.url_title}${(skill.skill_progress.level == 1)? "/practice":""}`;
+			skillLink.href = `/skill/${languageCode}/${skill.url_title}${(skill.skill_progress.level === 1)? "/practice":""}`;
 			skillLink.textContent = skill.short;
 		}
 
@@ -1875,16 +1875,16 @@ function displayNeedsStrengthening(needsStrengthening, cracked = false, needsSor
 	strengthenBox.removeChild(strengthenBox.lastChild);
 	
 	const listEndText = document.createTextNode("");
-	if (numSkillsToShow == 1)
+	if (numSkillsToShow === 1)
 		listEndText.textContent = ""; // If there is only one skill in the list, don't put anything before it.
-	else if (numSkillsToShow == numSkillsToBeStrengthened)
+	else if (numSkillsToShow === numSkillsToBeStrengthened)
 		listEndText.textContent = " & "; // Add & if we have put some stuff and showing every skill in list so the next one is the very last.
 	else
 		listEndText.textContent = ", "; // Otherwise we have put some stuff and there is more coming after so just put a comma.
 
 	strengthenBox.appendChild(listEndText);
 
-	if (numSkillsToShow == numSkillsToBeStrengthened)
+	if (numSkillsToShow === numSkillsToBeStrengthened)
 	{
 		const skillLink = document.createElement("a");
 		skillLink.addEventListener("focus", focus);
@@ -1901,7 +1901,7 @@ function displayNeedsStrengthening(needsStrengthening, cracked = false, needsSor
 			// last skill to be displayed is a bonus skill
 			const skill = needsStrengthening[1][needsStrengthening[1].length -1];
 
-			skillLink.href = `/skill/${languageCode}/${skill.url_title}${(skill.skill_progress.level == 1)? "/practice":""}`;
+			skillLink.href = `/skill/${languageCode}/${skill.url_title}${(skill.skill_progress.level === 1)? "/practice":""}`;
 			skillLink.textContent = skill.short;
 		} else
 		{
@@ -1947,7 +1947,7 @@ function displayNeedsStrengthening(needsStrengthening, cracked = false, needsSor
 			let bonusSkillIndex = lastIndexToBeShown - needsStrengthening[0].length;
 			const skill = needsStrengthening[1][bonusSkillIndex];
 
-			skillLink.href = `/skill/${languageCode}/${skill.url_title}${(skill.skill_progress.level == 1)? "/practice":""}`;
+			skillLink.href = `/skill/${languageCode}/${skill.url_title}${(skill.skill_progress.level === 1)? "/practice":""}`;
 			skillLink.textContent = skill.short;
 		}
 		strengthenBox.appendChild(skillLink);
@@ -2044,7 +2044,7 @@ function getSkillFromPopout(skillPopout)
 {
 	const skillTitle = skillPopout.parentNode.querySelector(SKILL_NAME_SELECTOR).textContent;
 	const allSkills = [...userData.language_data[languageCode].skills, ...userData.language_data[languageCode].bonus_skills];
-	return allSkills.find(skill => skill.short == skillTitle);
+	return allSkills.find(skill => skill.short === skillTitle);
 }
 
 function addSmallButtonsConatiner(skillPopout)
@@ -2134,7 +2134,7 @@ function addWordsButton(skillPopout)
 		(event) => {
 			const smallButtonsContainer = event.target.parentNode;
 			const wordsListBubble = smallButtonsContainer.querySelector("#wordsListBubble");
-			if (wordsListBubble == null)
+			if (wordsListBubble === null)
 				smallButtonsContainer.appendChild(createWordsListBubble(words, wordsButton, smallButtonsContainer, isLocked));
 			else
 				wordsListBubble.remove();
@@ -2243,11 +2243,11 @@ function addMasteredSkillButton(skillPopout)
 
 function addPractiseButton(skillPopout)
 {
-	if (skillPopout == null)
+	if (skillPopout === null)
 		return false;
 	
 	const levelContainer = document.querySelector(SKILL_POPOUT_LEVEL_CONTAINER_SELECTOR);
-	if (levelContainer == null)
+	if (levelContainer === null)
 		return false; // Locked skill so don't do anything
 
 	const skillLevel = levelContainer.textContent.slice(-3,-2);
@@ -3153,7 +3153,7 @@ function displayCrownsBreakdown()
 				{
 					const skillCount = crownLevelCount[skillType][crownLevel];
 
-					if (!options.crownsBreakdownShowZerosRows && skillCount == 0)
+					if (!options.crownsBreakdownShowZerosRows && skillCount === 0)
 					{
 						continue;
 					}
@@ -3172,7 +3172,7 @@ function displayCrownsBreakdown()
 
 					const skillsAtSpan = document.createElement("span");
 					skillsAtSpan.classList.add("skillsAtSpan");
-					skillsAtSpan.textContent = `skill${skillCount == 1 ? "" : "s"} at`;
+					skillsAtSpan.textContent = `skill${skillCount === 1 ? "" : "s"} at`;
 					breakdownListItem.appendChild(skillsAtSpan);
 
 					breakdownListItem.appendChild(imgContainer.cloneNode(true));
@@ -3184,7 +3184,7 @@ function displayCrownsBreakdown()
 					breakdownListItem.appendChild(crownCountSpan);
 
 					const crownsSpan = document.createElement("span");
-					crownsSpan.textContent = `crown${(crownCount == 1 )?"":"s"}`;
+					crownsSpan.textContent = `crown${(crownCount === 1 )?"":"s"}`;
 					breakdownListItem.appendChild(crownsSpan);
 
 					breakdownList.appendChild(breakdownListItem);
@@ -3196,7 +3196,7 @@ function displayCrownsBreakdown()
 			if (options.crownsBreakdown) crownsInfoContainer.appendChild(breakdownContainer);
 
 			// Checkpoint Prediction
-			if (treeLevel == 0 && options.checkpointPrediction)
+			if (treeLevel === 0 && options.checkpointPrediction)
 			{
 				const predictionData = (progress.length > 5) ? daysToNextCheckpoint() : daysToNextCheckpointByCalendar();
 
@@ -3355,7 +3355,7 @@ function displayXPBreakdown()
 function displayTotalStrenthBox()
 {
 	const sidebar = document.querySelector(`.${SIDEBAR}`);
-	if (sidebar == null) return false;
+	if (sidebar === null) return false;
 
 	const skills = userData.language_data[languageCode].skills.filter(skill => skill.skill_progress.level !== 0);
 	if (skills.length === 0) return false;
@@ -3417,9 +3417,9 @@ function displayTotalStrenthBox()
 function displayLanguagesInfo(languages)
 {
 	const sidebar = document.querySelector(`.${SIDEBAR}`);
-	if (sidebar == null) return false;
+	if (sidebar === null) return false;
 
-	if (languages.length == 0)
+	if (languages.length === 0)
 	{
 		removeLanguagesInfo();
 		return false;
@@ -3478,7 +3478,7 @@ function displayLanguagesInfo(languages)
 
 	// Add the new side bar box to the page
 	const dailyGoalBox = sidebar.querySelector(`.${DAILY_GOAL_SIDEBAR_CONTAINER}`);
-	if (dailyGoalBox == null)
+	if (dailyGoalBox === null)
 	{
 		return false;
 	}
@@ -3513,7 +3513,7 @@ function displaySuggestion(fullyStrengthened, noCrackedSkills)
 	else
 	{
 		// body hasn't loaded yet so element not there, let's try again after a small wait, but only if we are still on the main page.
-		if(onMainPage && document.querySelector(`[data-test="intro-lesson"]`) == null)
+		if(onMainPage && document.querySelector(`[data-test="intro-lesson"]`) === null)
 		{
 			setTimeout(function(){displaySuggestion(fullyStrengthened, noCrackedSkills);}, 50);
 		}
@@ -3637,7 +3637,7 @@ function displaySuggestion(fullyStrengthened, noCrackedSkills)
 				const checkpoints = document.querySelectorAll(CHECKPOINT_SELECTOR);
 				checkpoints.forEach(
 					(checkpoint, index) => {
-						if (checkpointNumber == null && checkpoint.querySelector(`img`).src.includes(`unlocked`))
+						if (checkpointNumber === null && checkpoint.querySelector(`img`).src.includes(`unlocked`))
 							checkpointNumber = index;
 					}
 				);
@@ -4446,7 +4446,7 @@ function httpGetAsync(url, responseHandler)
     let xmlHttp = new XMLHttpRequest();
     xmlHttp.onreadystatechange = function()
 	{ 
-        if (xmlHttp.readyState == 4 && xmlHttp.status == 200)
+        if (xmlHttp.readyState === 4 && xmlHttp.status === 200)
             responseHandler(xmlHttp.responseText);
     };
     xmlHttp.open("GET", url, true); // true for asynchronous 
@@ -4466,9 +4466,9 @@ function httpGetAsync(url, responseHandler)
 			let xmlHttp = new XMLHttpRequest();
 			xmlHttp.onreadystatechange = function()
 			{
-				if (xmlHttp.readyState == 4)
+				if (xmlHttp.readyState === 4)
 				{
-					if (xmlHttp.status == 200)
+					if (xmlHttp.status === 200)
 					{
 						document.getElementById('userData${requestID}').textContent = "//" + xmlHttp.responseText;		
 					}
@@ -4509,7 +4509,7 @@ function requestResponseMutationHandle(mutationsList, url, responseHandler)
 
 		--requestsPending; // We have recieved some sort of response for this request.
 
-		if (dataElem.textContent.slice(0,7) == "//ERROR")
+		if (dataElem.textContent.slice(0,7) === "//ERROR")
 		{
 			// The request had an error, lets clear the scripts and try again after a short wait.
 			const code = dataElem.textContent.slice(8);
@@ -4567,7 +4567,7 @@ async function handleDataResponse(responseText)
 	if (languageChanged)
 	{
 		// The language change hasn't been resolved yet.
-		if (newDataLanguageCode + newDataUICode == languageCode + UICode)
+		if (newDataLanguageCode + newDataUICode === languageCode + UICode)
 		{
 			// language change has happened but the data isn't up to date yet as it is not matching the current active language
 			// so request the data, but only if still on the main page. Safe to not wait as the httpRequest will take some time.
@@ -4755,12 +4755,12 @@ function revealNewWord()
 
 function hideTranslationText(reveal = false, setupObserver = true)
 {
-	if (document.getElementsByClassName(QUESTION_CONTAINER).length == 0)
+	if (document.getElementsByClassName(QUESTION_CONTAINER).length === 0)
 		return false;
 	
 	const questionContainer = document.querySelector(`.${QUESTION_CONTAINER}`);
 
-	if (questionContainer.firstChild == null)
+	if (questionContainer.firstChild === null)
 		return false; // Duo ecouragement message, no question box, do nothing.
 
 	const questionTypeString = questionContainer.firstChild.getAttribute("data-test");
@@ -4813,7 +4813,7 @@ function hideTranslationText(reveal = false, setupObserver = true)
 						const hotkeyList = options.revealHotkeyCode.split("+");
 						const numKeys = hotkeyList.length;
 						if (
-							e.key.toUpperCase() == hotkeyList[numKeys-1]
+							e.key.toUpperCase() === hotkeyList[numKeys-1]
 							&& ( (hotkeyList.includes("Ctrl") && e.ctrlKey) || !hotkeyList.includes("Ctrl") )
 							&& ( (hotkeyList.includes("Shift") && e.shiftKey) || !hotkeyList.includes("Shift") )
 							&& ( (hotkeyList.includes("Meta") && e.metaKey) || !hotkeyList.includes("Meta") )
@@ -4927,6 +4927,7 @@ function childListMutationHandle(mutationsList, observer)
 	let rootChildContentsReplaced = false;
 	let rootChildRemovedNodes = [];
 	let mainBodyReplaced = false
+	let topBarToggled = false;
 	let bottomNavToggled = false;
 	let popupChanged = false;
 	let popupIcon;
@@ -4961,22 +4962,37 @@ function childListMutationHandle(mutationsList, observer)
 		{
 			goldenOwlMessageAdded = true;
 		}
-		else if (mutation.target == rootChild)
+		else if (mutation.target === rootChild)
 		{
 			rootChildContentsReplaced = true;
 			rootChildRemovedNodes = rootChildRemovedNodes.concat(Array.from(mutation.removedNodes));
 		}
-		else if (mutation.target == mainBodyContainer)
+		else if (mutation.target === mainBodyContainer)
 		{
 			mainBodyReplaced = true;
 		}
-		else if (
+		else if
+		(
 			(
 				[
 					...Array.from(mutation.addedNodes),
 					...Array.from(mutation.removedNodes)
 				].some(
-					(addedNode) => addedNode.className.includes(BOTTOM_NAV_SELECTOR.substring(1))
+					(changedNode) => (changedNode.querySelector(`.${TOP_BAR}`) !== null || changedNode.matches(`.${TOP_BAR}`))
+				) // The top bar was added or removed
+			)
+		)
+		{
+			topBarToggled = true;
+		}
+		else if
+		(
+			(
+				[
+					...Array.from(mutation.addedNodes),
+					...Array.from(mutation.removedNodes)
+				].some(
+					(changedNode) => (changedNode.querySelector(BOTTOM_NAV_SELECTOR) !== null || changedNode.matches(BOTTOM_NAV_SELECTOR))
 				) // The bottom nav was added or removed
 			)
 		)
@@ -4984,7 +5000,7 @@ function childListMutationHandle(mutationsList, observer)
 			bottomNavToggled = true;
 		}
 		else if (
-			mutation.target.className == POPUP_ICON &&
+			mutation.target.className === POPUP_ICON &&
 			mutation.addedNodes.length != 0
 		)
 		{
@@ -4993,7 +5009,7 @@ function childListMutationHandle(mutationsList, observer)
 		}
 		else if (
 			mutation.target.className.includes(LESSON_MAIN_SECTION)
-			&& Array.from(mutation.removedNodes).some(node => `.${node.className}` === PRACTICE_TYPE_SELECT_MESSAGE_SELECTOR)
+			&& Array.from(mutation.removedNodes).some(node => node.matches(PRACTICE_TYPE_SELECT_MESSAGE_SELECTOR))
 		)
 		{
 			lessonMainSectionContentsReplaced = true;
@@ -5003,16 +5019,16 @@ function childListMutationHandle(mutationsList, observer)
 			lessonQuestionChanged = true;
 		}
 		else if (
-			mutation.target.attributes.hasOwnProperty("data-test") &&
-			mutation.target.attributes["data-test"].value == "skill-icon" &&
-			mutation.removedNodes.length == 1 &&
-			`.${mutation.removedNodes[0].className}` == CRACKED_SKILL_OVERLAY_SELECTOR
+			mutation.target.attributes.hasOwnProperty("data-test")
+			&& mutation.target.attributes["data-test"].value === "skill-icon"
+			&& mutation.removedNodes.length === 1
+			&& mutation.removedNodes[0].matches(CRACKED_SKILL_OVERLAY_SELECTOR)
 		)
 		{
 			skillRepaired = true;
 		}
 		else if (
-			mutation.target.getAttribute("data-test") == "skill" &&
+			mutation.target.getAttribute("data-test") === "skill" &&
 			mutation.addedNodes.length != 0 &&
 			mutation.target.querySelectorAll(`[data-test="skill-popout"]`).length !== 0
 		)
@@ -5020,10 +5036,11 @@ function childListMutationHandle(mutationsList, observer)
 			skillPopoutAdded = true;
 			skillPopout = mutation.target.querySelector(`[data-test="skill-popout"]`);
 		}
-		else if (
-			`.${mutation.target.className}` == CHECKPOINT_CONTAINER_SELECTOR &&
-			mutation.addedNodes.length != 0 &&
-			mutation.target.querySelector(CHECKPOINT_POPOUT_SELECTOR) != null
+		else if
+		(
+			mutation.target.matches(CHECKPOINT_CONTAINER_SELECTOR)
+			&& mutation.addedNodes.length !== 0
+			&& mutation.target.querySelector(CHECKPOINT_POPOUT_SELECTOR) !== null
 		)
 		{
 			checkpointPopoutAdded = true;
@@ -5039,13 +5056,13 @@ function childListMutationHandle(mutationsList, observer)
 	else if (rootChildContentsReplaced)
 	{
 		// Check if there is both the topbar and the main page elem.
-		if (rootChild.childElementCount == 2)
+		if (rootChild.childElementCount === 2)
 		{
 			// not just changed into a lesson
 			languageChanged = false;
 			init();
 		}
-		else if (rootChild.childElementCount == 1)
+		else if (rootChild.childElementCount === 1)
 		{
 			// Entered a lesson in the normal way, through the skill tree.
 			// Don't need to do anything here as when the lesson will be loading first.
@@ -5089,6 +5106,12 @@ function childListMutationHandle(mutationsList, observer)
 		}
 	}
 
+	if (topBarToggled && (new RegExp("/tips/?")).test(window.location.pathname))
+	{
+		// Swithced between mobile and desktop layouts on a tips page.
+		init();
+	}
+
 	if (bottomNavToggled)
 	{
 		// Switched between mobile and desktop layouts.
@@ -5130,12 +5153,6 @@ function childListMutationHandle(mutationsList, observer)
 			}
 		);
 
-		if (document.querySelector(MOBILE_TIPS_PAGE_HEADER_SELECTOR) === null)
-		{
-			// Re set up the observers as topBarDiv will have been replaced and there might be a bottomNav
-			setUpObservers();
-		}
-
 		// Try to re apply focus mode option
 		applyFocusMode();
 
@@ -5165,7 +5182,7 @@ function childListMutationHandle(mutationsList, observer)
 		if (popupIcon.querySelector(`.${GOLD_CROWN}`) !== null) // Grey crown doesn't seem to have its own class now, it has the same as gold.
 		{
 			// Crowns has had the change.
-			if (options.crownsInfo && popupIcon.lastChild.nodeName == 'DIV')
+			if (options.crownsInfo && popupIcon.lastChild.nodeName === 'DIV')
 			{
 				// Pop-up box, which is a div, has just appeared as the last child, let's display the Crown breakdown.
 				displayCrownsBreakdown();
@@ -5184,7 +5201,7 @@ function childListMutationHandle(mutationsList, observer)
 		) // Lit flame for streak extended today, grey for not, blue possibly for frozen with duolingo plus?
 		{
 			// Streak/XP has had the change.
-			if (options.XPInfo && popupIcon.lastChild.nodeName == 'DIV')
+			if (options.XPInfo && popupIcon.lastChild.nodeName === 'DIV')
 			{
 				// Pop-up box, which is a div, has just appeared as the last child, let's display the XP breakdown.
 				displayXPBreakdown();
@@ -5230,7 +5247,7 @@ function childListMutationHandle(mutationsList, observer)
 	if (skillPopoutAdded)
 	{
 		const introLesson = document.querySelector(`[data-test="intro-lesson"]`);
-		if (introLesson == null || !introLesson.contains(skillPopout))
+		if (introLesson === null || !introLesson.contains(skillPopout))
 		{
 			// No introduction lesson, or if there is this skillPopout isn't from that skill
 			if (options.practiseButton) addPractiseButton(skillPopout);
@@ -5264,12 +5281,12 @@ function classNameMutationHandle(mutationsList, observer)
 	let questionCheckStatusChange = false;
 	for (let mutation of mutationsList)
 	{
-		if (mutation.target.parentNode.parentNode == languageLogo)
+		if (mutation.target.parentNode.parentNode === languageLogo)
 		{
 			// it was a language change
 			isLanguageChange = true;
 		}
-		else if (mutation.target.parentNode.className == LESSON_BOTTOM_SECTION)
+		else if (mutation.target.parentNode.className === LESSON_BOTTOM_SECTION)
 		{
 			// The question check status changed
 			questionCheckStatusChange = true;
@@ -5307,7 +5324,7 @@ function classNameMutationHandle(mutationsList, observer)
 		applyFixedSidebar();
 
 		// check if we are now on the main page
-		if (window.location.pathname == "/learn")
+		if (window.location.pathname === "/learn")
 		{
 			// on main page
 			onMainPage = true;
@@ -5389,7 +5406,7 @@ function setUpObservers()
 
 		homeNav = topBarDiv.childNodes[0];
 
-		if (numNavButtons == 5)
+		if (numNavButtons === 5)
 		{
 			storiesNav = topBarDiv.childNodes[2];
 			/* unused/unusable
@@ -5469,7 +5486,7 @@ async function init()
 	rootElem = document.getElementById("root"); // When logging in child list is changed.
 	childListObserver.observe(rootElem,{childList: true}); // Observing for changes to its children to detect logging in and out?
 
-	if (rootElem.childElementCount == 0)
+	if (rootElem.childElementCount === 0)
 		return false;
 
 	rootChild = rootElem.childNodes[0];
@@ -5485,13 +5502,13 @@ async function init()
 	{
 		// On the courses page, here the main body container is the first child of rootChild for some reason.
 		mainBodyContainer = rootChild.firstChild;
-		if (mainBodyContainer == null)
+		if (mainBodyContainer === null)
 			return false;
 	}
 	else
 	{
 		mainBodyContainer = rootChild.lastChild;
-		if (mainBodyContainer == null)
+		if (mainBodyContainer === null)
 			return false;
 
 	}
@@ -5500,7 +5517,7 @@ async function init()
 
 
 	mainBody = mainBodyContainer.firstChild;
-	if (mainBody == null)
+	if (mainBody === null)
 		return false;
 	
 	if (
@@ -5511,6 +5528,15 @@ async function init()
 		inMobileLayout = true;
 		rootElem.classList.add("mobileLayout");
 		rootElem.classList.remove("desktopLayout");
+		if (
+			(new RegExp("/tips/?")).test(window.location.pathname)
+			&& document.querySelector(`.${TOP_BAR}`) === null
+			&& document.querySelector(NAVS_CONTAINER_SELECTOR) !== null
+		)
+		{
+			// Set up observer for the container of the nav bars when in mobile layout on tips page so we can detect if we change out of it.
+			childListObserver.observe(document.querySelector(NAVS_CONTAINER_SELECTOR), {childList: true});
+		}
 	}
 	else
 	{
@@ -5533,7 +5559,7 @@ async function init()
 
 		// now test to see if we are in a lesson or not
 
-		if (rootChild.firstChild.className == LESSON)
+		if (rootChild.firstChild.className === LESSON)
 		{
 			// in a lesson
 			// we probably got here from a link in the needs strengthening list or from a practiseButton
@@ -5632,17 +5658,12 @@ async function init()
 		else
 		{
 			// not in a lesson
-			/*
-				Main body container element has class _3MLiB.
-				If it is the first child, there is no topBarDiv and it is the only child of rootChild,
-				if it is second place, then the first child should be a topBarDiv and there are two children of rootChild.
-				So it is enough to just test for the number of children of rootChild to check for a topBarDiv.
-			*/
 
 			questionNumber = 1;
-			if (rootChild.childElementCount === 1 || document.querySelector(MOBILE_TIPS_PAGE_HEADER_SELECTOR) !== null)
+			if (document.querySelector(`.${TOP_BAR}`) === null)
 			{
 				// no topBarDiv
+				// set up the observer to check for layout changes where the bottomNav might be added or removed
 				onMainPage = false;
 				return false;
 			}
@@ -5690,7 +5711,7 @@ async function init()
 
 				const popout = document.querySelector(`[data-test="skill-popout"], ${CHECKPOINT_POPOUT_SELECTOR}`);
 
-				if (popout != null)
+				if (popout !== null)
 					popout.scrollIntoView({block: "center"});
 
 				// Done all the prep we need, let's get some data to process